feat: Add findManyByUserId method to CheckInsRepository
This commit adds a new method, `findManyByUserId`, to the `CheckInsRepository` interface. This method allows retrieving multiple check-ins by user ID and page number. The implementation of this method is also added to the `InMemoryCheckInsRepository` class.

The purpose of this change is to enable fetching paginated check-in history for a specific user.